Moved ColorPickerDialog to androidshared module
