+++ androidshared/build.gradle
@@ -6,6 +6,10 @@ plugins {
 android {
     compileSdkVersion rootProject.androidCompileSdk
 
+    buildFeatures {
+        viewBinding true
+    }
+
     defaultConfig {
         minSdkVersion rootProject.androidMinSdk
         targetSdkVersion rootProject.androidTargetSdk
@@ -35,8 +39,12 @@ dependencies {
     implementation 'androidx.core:core-ktx:1.3.2'
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.3.0"
     implementation "com.google.android.material:material:${rootProject.materialVersion}"
+    implementation "androidx.fragment:fragment-ktx:1.3.4"
     implementation project(':strings')
 
+    testImplementation project(':testshared')
     testImplementation "junit:junit:4.13.2"
+    testImplementation 'androidx.test.ext:junit:1.1.2'
+    testImplementation 'androidx.test.espresso:espresso-core:3.3.0'
     testImplementation "org.robolectric:robolectric:4.5.1"
 }
