+++ androidshared/src/main/res/anim/popup_hide.xml
