+++ androidshared/src/main/res/color/color_on_surface_high_emphasis.xml
