+++ androidshared/src/main/res/color/light_theme_tint_list.xml
