+++ androidshared/src/main/res/color/map_button_tint_list.xml
