+++ androidshared/src/main/res/drawable/color_circle.xml
