+++ androidshared/src/main/res/drawable/thumb.xml
