+++ androidshared/src/main/res/layout/color_picker_dialog_layout.xml
