+++ androidshared/src/main/res/values/attrs.xml
