+++ androidshared/src/main/res/values/buttons.xml
