+++ androidshared/src/main/res/values/colors.xml
