+++ androidshared/src/main/res/values/fonts.xml
