+++ androidshared/src/main/res/values/light_theme.xml
