+++ androidshared/src/main/res/values/magenta_theme.xml
