+++ androidshared/src/main/res/values/styles.xml
